// client-service/app/pages/profileAddPage/enums.ts
export enum EProfileAddFormFields {
  DisplayName = "displayName",
  Username = "userName",
  Email = "email",
  MobileNumber = "mobileNumber",
  Password = "password",
  PasswordConfirm = "passwordConfirm",
  Birthday = "birthday",
  Gender = "gender",
  SearchGender = "searchGender",
  Location = "location",
  Description = "description",
  Height = "height",
  Weight = "weight",
  LookingFor = "lookingFor",
  Image = "image",
  TelegramID = "telegramId",
  TelegramUsername = "telegramUserName",
  FirstName = "firstName",
  LastName = "lastName",
  LanguageCode = "languageCode",
  AllowsWriteToPm = "allowsWriteToPm",
  QueryId = "queryId",
  Latitude = "latitude",
  Longitude = "longitude",
  AgeFrom = "ageFrom",
  AgeTo = "ageTo",
  Distance = "distance",
  Page = "page",
  Size = "size",
}